refactor(default_chatter): 合并 pass_and_wait 和 stop_conversation 到 control 动作模块
