+++ plugins/default_chatter/actions/__init__.py
@@ -2,9 +2,8 @@
 
 from __future__ import annotations
 
-from .pass_and_wait import PassAndWaitAction
+from .control import PassAndWaitAction, StopConversationAction
 from .send_text import SendTextAction
-from .stop_conversation import StopConversationAction
 
 __all__ = [
     "SendTextAction",
